Restrict charge tables to active bookings only
Problem
-------

The charge category and charge type tables
display counts for all charges in the database.

The rest of the page limits results to only active bookings,
so these tables display extra information that is out of context.

Solution
--------

Use [`ActiveRecord#merge`] to limit displayed charges to only those
belonging to active bookings.

[`ActiveRecord#merge`]: http://apidock.com/rails/ActiveRecord/SpawnMethods/merge