[Refactor] Keep track of which filters are applied
Keep a filters data structure in the application JS, with the structure:

```json
{
  "status": {
    "Pre-trial": false,
    "Sentenced": false
  },
  "location":{
    "Main Jail Complex": false,
    "County Correctional Center": false,
    "Medical Facility": false
  },
  "gender":{
    "male": false,
    "female": false
  },
  "race":{
    "American Indian": false,
    "Hispanic": false,
    "White": false,
    "Black or African American": false,
    "Asian": false,
    "Middle Eastern": false,
    "Alaska Native": false
  }
}
```

The JS `FilterTable` object stores this data structure as `this.filters`,
and mutates the boolean values as filters are applied and removed.

When `update()` is called, the object clears previous filters,
applies the new relevant filters,
and re-renders the object.

There are a few gotchas:

* Only one filter can be applied to a dimension at a time
* If more than one filter is applied,
  the last one in the dimension will be the only one to take effect.
* There are still no indications of which filters are applied -
  this is only an internal refactoring commit.